requirements.txt
openpyxl==3.1.3
packaging==24.1
tqdm==4.66.4
pyyaml==6.0.1